refactor/move IFile interface from model to interfaces IFile.ts
