+++ server/Interfaces/IFile.ts
@@ -1,0 +1,8 @@
+import { Document } from "mongoose";
+
+export interface IFile extends Document {
+   filename: string;
+   filedata: string;
+   createdAt?: Date;
+   updatedAt?: Date;
+}
