+++ server/models/fileModel.ts
@@ -1,11 +1,5 @@
-import mongoose, { Document, Schema, Model } from "mongoose";
-
-export interface IFile extends Document {
-   filename: string;
-   filedata: string;
-   createdAt?: Date;
-   updatedAt?: Date;
-}
+import mongoose, { Schema, Model } from "mongoose";
+import { IFile } from "../Interfaces/IFile";
 
 const fileSchema: Schema<IFile> = new Schema(
    {
